:recycle: api: extract controller logic to usecase for completed v3
